+++ raiden-ts/src/actions.ts
@@ -28,10 +28,7 @@ export const raidenShutdown = createAction(
 );
 export interface raidenShutdown extends ActionType<typeof raidenShutdown> {}
 
-export const raidenConfigUpdate = createAction(
-  'raidenConfigUpdate',
-  t.type({ config: PartialRaidenConfig }),
-);
+export const raidenConfigUpdate = createAction('raidenConfigUpdate', PartialRaidenConfig);
 export interface raidenConfigUpdate extends ActionType<typeof raidenConfigUpdate> {}
 
 const RaidenActions = {
@@ -79,6 +76,7 @@ export const ConfirmableActions = [
   ChannelsActions.channelWithdrawn,
   ChannelsActions.channelClose.success,
   ChannelsActions.channelSettle.success,
+  TransfersActions.transferSecretRegistered,
 ];
 /**
  * Union of codecs of actions above
@@ -89,10 +87,12 @@ export const ConfirmableAction = t.union([
   ChannelsActions.channelWithdrawn.codec,
   ChannelsActions.channelClose.success.codec,
   ChannelsActions.channelSettle.success.codec,
+  TransfersActions.transferSecretRegistered.codec,
 ]);
 export type ConfirmableAction =
   | ChannelsActions.channelOpen.success
   | ChannelsActions.channelDeposit.success
   | ChannelsActions.channelWithdrawn
   | ChannelsActions.channelClose.success
-  | ChannelsActions.channelSettle.success;
+  | ChannelsActions.channelSettle.success
+  | TransfersActions.transferSecretRegistered;