+++ raiden-ts/src/transfers/epics.ts
@@ -26,13 +26,14 @@ import {
   withLatestFrom,
 } from 'rxjs/operators';
 import { bigNumberify } from 'ethers/utils';
+import { Event } from 'ethers/contract';
 import { One, Zero } from 'ethers/constants';
 import { findKey, get, pick, isMatchWith } from 'lodash';
 
 import { RaidenEpicDeps } from '../types';
 import { RaidenAction } from '../actions';
 import { RaidenState } from '../state';
-import { Address, assert, Hash, Signed, UInt, BigNumberC } from '../utils/types';
+import { Address, assert, Hash, Signed, UInt, BigNumberC, Secret } from '../utils/types';
 import { isActionOf, isResponseOf } from '../utils/actions';
 import { LruCache } from '../utils/lru';
 import { messageReceived, messageSend } from '../messages/actions';
@@ -58,6 +59,7 @@ import { RaidenConfig } from '../config';
 import { matrixPresence } from '../transport/actions';
 import { pluckDistinct } from '../utils/rx';
 import { RaidenError, ErrorCodes } from '../utils/error';
+import { getEventsStream } from '../utils/ethers';
 import {
   transfer,
   transferExpire,
@@ -65,6 +67,7 @@ import {
   transferProcessed,
   transferRefunded,
   transferSecret,
+  transferSecretRegistered,
   transferSecretRequest,
   transferSecretReveal,
   transferSigned,
@@ -100,15 +103,15 @@ function dispatchAndWait$<A extends RaidenAction>(
   predicate: (action: RaidenAction) => boolean,
 ): Observable<A> {
   return merge(
-    // output once
-    of(request),
-    // but wait until respective success/failure action is seen before completing
+    // wait until respective success/failure action is seen before completing
     action$.pipe(
       filter(predicate),
       take(1),
       // don't output success/failure action, just wait for first match to complete
       ignoreElements(),
     ),
+    // output once
+    of(request),
   );
 }
 
@@ -119,7 +122,6 @@ function retryUntil<T>(notifier: Observable<any>, delayMs = 30e3): MonoTypeOpera
   // waits for address's user in transport to be online and joined room before actually
   // sending the message. That's why repeatWhen emits/resubscribe only some time after
   // sendOnceAndWaitSent$ completes, instead of a plain 'interval'
-  // TODO: configurable retry delay, possibly use an exponential backoff timeout strat
   return input$ =>
     input$.pipe(
       repeatWhen(completed$ => completed$.pipe(delay(delayMs))),
@@ -132,9 +134,10 @@ function retrySendUntil$(
   action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
   predicate: (state: RaidenState) => boolean,
+  delayMs = 30e3,
 ): Observable<messageSend.request> {
   return dispatchAndWait$(action$, send, isResponseOf(messageSend, send.meta)).pipe(
-    retryUntil(state$.pipe(filter(predicate))),
+    retryUntil(state$.pipe(filter(predicate)), delayMs),
   );
 }
 
@@ -226,11 +229,11 @@ function makeAndSignTransfer$(
   };
   return from(signMessage(signer, message, { log })).pipe(
     mergeMap(function*(signed) {
+      // messageSend LockedTransfer handled by transferSignedRetryMessageEpic
+      yield transferSigned({ message: signed, fee }, action.meta);
       // besides transferSigned, also yield transferSecret (for registering) if we know it
       if (action.payload.secret)
         yield transferSecret({ secret: action.payload.secret }, action.meta);
-      yield transferSigned({ message: signed, fee }, action.meta);
-      // messageSend LockedTransfer handled by transferSignedRetryMessageEpic
     }),
   );
 }
@@ -292,7 +295,12 @@ function makeAndSignUnlock$(
     signed$ = of(state.sent[secrethash].unlock![1]);
   } else {
     // don't forget to check after signature too, may have expired by then
-    assert(transfer.lock.expiration.gt(state.blockNumber), 'lock expired');
+    // allow unlocking past expiration if secret registered
+    assert(
+      state.sent[secrethash].secret?.[1]?.registerBlock ||
+        transfer.lock.expiration.gt(state.blockNumber),
+      'lock expired',
+    );
     const locksroot = getChannelLocksroot(channel, secrethash);
 
     const message: Unlock = {
@@ -308,7 +316,7 @@ function makeAndSignUnlock$(
       locked_amount: channel.own.balanceProof.lockedAmount.sub(transfer.lock.amount) as UInt<32>,
       locksroot,
       payment_identifier: transfer.payment_identifier,
-      secret: state.secrets[action.meta.secrethash].secret,
+      secret: state.sent[action.meta.secrethash].secret![1].value,
     };
     signed$ = from(signMessage(signer, message, { log }));
   }
@@ -316,7 +324,11 @@ function makeAndSignUnlock$(
   return signed$.pipe(
     withLatestFrom(state$),
     mergeMap(function*([signed, state]) {
-      assert(transfer.lock.expiration.gt(state.blockNumber), 'lock expired!');
+      assert(
+        state.sent[secrethash].secret?.[1]?.registerBlock ||
+          transfer.lock.expiration.gt(state.blockNumber),
+        'lock expired',
+      );
       assert(!state.sent[secrethash].channelClosed, 'channel closed!');
       yield transferUnlock.success({ message: signed }, action.meta);
       // messageSend Unlock handled by transferUnlockedRetryMessageEpic
@@ -611,6 +623,7 @@ const transferSignedRetryMessage$ = (
   action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
   action: transferSigned,
+  { httpTimeout }: RaidenConfig,
 ): Observable<messageSend.request> => {
   const secrethash = action.meta.secrethash;
   const signed = action.payload.message;
@@ -628,7 +641,7 @@ const transferSignedRetryMessage$ = (
       !!transfer.channelClosed
     );
   };
-  return retrySendUntil$(send, action$, state$, processedOrNotPossibleToSend);
+  return retrySendUntil$(send, action$, state$, processedOrNotPossibleToSend, httpTimeout);
 };
 
 /**
@@ -639,18 +652,19 @@ const transferSignedRetryMessage$ = (
  *
  * @param action$ - Observable of transferSigned actions
  * @param state$ - Observable of RaidenStates
- * @param latest$ - RaidenEpicDeps latest
+ * @param deps - RaidenEpicDeps
  * @returns Observable of messageSend.request actions
  */
 export const transferSignedRetryMessageEpic = (
   action$: Observable<RaidenAction>,
   {}: Observable<RaidenState>,
-  { latest$ }: RaidenEpicDeps,
+  { latest$, config$ }: RaidenEpicDeps,
 ): Observable<messageSend.request> =>
   action$.pipe(
     filter(isActionOf(transferSigned)),
-    mergeMap(action =>
-      transferSignedRetryMessage$(action$, latest$.pipe(pluckDistinct('state')), action),
+    withLatestFrom(config$),
+    mergeMap(([action, config]) =>
+      transferSignedRetryMessage$(action$, latest$.pipe(pluckDistinct('state')), action, config),
     ),
   );
 
@@ -668,6 +682,7 @@ const transferUnlockedRetryMessage$ = (
   state$: Observable<RaidenState>,
   action: transferUnlock.success,
   state: RaidenState,
+  { httpTimeout }: RaidenConfig,
 ): Observable<messageSend.request> => {
   const secrethash = action.meta.secrethash;
   if (!(secrethash in state.sent)) return EMPTY; // shouldn't happen
@@ -684,7 +699,7 @@ const transferUnlockedRetryMessage$ = (
     return !!transfer.unlockProcessed || !!transfer.channelClosed;
   };
 
-  return retrySendUntil$(send, action$, state$, unlockProcessedOrChannelClosed);
+  return retrySendUntil$(send, action$, state$, unlockProcessedOrChannelClosed, httpTimeout);
 };
 
 /**
@@ -700,13 +715,19 @@ const transferUnlockedRetryMessage$ = (
 export const transferUnlockedRetryMessageEpic = (
   action$: Observable<RaidenAction>,
   {}: Observable<RaidenState>,
-  { latest$ }: RaidenEpicDeps,
+  { latest$, config$ }: RaidenEpicDeps,
 ): Observable<messageSend.request> =>
   action$.pipe(
     filter(isActionOf(transferUnlock.success)),
-    withLatestFrom(latest$.pipe(pluckDistinct('state'))),
-    mergeMap(([action, state]) =>
-      transferUnlockedRetryMessage$(action$, latest$.pipe(pluckDistinct('state')), action, state),
+    withLatestFrom(latest$.pipe(pluckDistinct('state')), config$),
+    mergeMap(([action, state, config]) =>
+      transferUnlockedRetryMessage$(
+        action$,
+        latest$.pipe(pluckDistinct('state')),
+        action,
+        state,
+        config,
+      ),
     ),
   );
 
@@ -724,6 +745,7 @@ const expiredRetryMessages$ = (
   state$: Observable<RaidenState>,
   action: transferExpire.success,
   state: RaidenState,
+  { httpTimeout }: RaidenConfig,
 ): Observable<messageSend.request> => {
   const secrethash = action.meta.secrethash;
   if (!(secrethash in state.sent)) return EMPTY; // shouldn't happen
@@ -740,7 +762,7 @@ const expiredRetryMessages$ = (
     return !!transfer.lockExpiredProcessed || !!transfer.channelClosed;
   };
   // emit request once immediatelly, then wait until respective success, then retries until confirmed
-  return retrySendUntil$(send, action$, state$, lockExpiredProcessedOrChannelClosed);
+  return retrySendUntil$(send, action$, state$, lockExpiredProcessedOrChannelClosed, httpTimeout);
 };
 
 /**
@@ -757,28 +779,30 @@ const expiredRetryMessages$ = (
 export const transferExpiredRetryMessageEpic = (
   action$: Observable<RaidenAction>,
   {}: Observable<RaidenState>,
-  { latest$ }: RaidenEpicDeps,
+  { latest$, config$ }: RaidenEpicDeps,
 ): Observable<messageSend.request> =>
   action$.pipe(
     filter(isActionOf(transferExpire.success)),
-    withLatestFrom(latest$.pipe(pluckDistinct('state'))),
-    mergeMap(([action, state]) =>
-      expiredRetryMessages$(action$, latest$.pipe(pluckDistinct('state')), action, state),
+    withLatestFrom(latest$.pipe(pluckDistinct('state')), config$),
+    mergeMap(([action, state, config]) =>
+      expiredRetryMessages$(action$, latest$.pipe(pluckDistinct('state')), action, state, config),
     ),
   );
 
 /**
  * Contains the core logic of {@link transferAutoExpireEpic}.
  *
- * @param state - Contains The current state of the app
- * @param blockNumber - The current block number
  * @param action$ - Observable of {@link RaidenAction} actions
+ * @param state - Contains The current state of the app
+ * @param config - Contains the current app config
+ * @param blockNumber - The current block number
  * @returns Observable of {@link transferExpire.request} or {@link transfer.failure} actions
  */
 function autoExpire$(
-  state: RaidenState,
-  blockNumber: number,
   action$: Observable<RaidenAction>,
+  state: RaidenState,
+  { confirmationBlocks }: RaidenConfig,
+  blockNumber: number,
 ): Observable<transferExpire.request | transfer.failure> {
   const requests$: Observable<transferExpire.request | transfer.failure>[] = [];
 
@@ -787,7 +811,10 @@ function autoExpire$(
       sent.unlock ||
       sent.lockExpired ||
       sent.channelClosed ||
-      sent.transfer[1].lock.expiration.gte(blockNumber)
+      sent.transfer[1].lock.expiration.add(confirmationBlocks).gt(blockNumber) ||
+      // don't expire if secret got registered before lock expired
+      (sent.secret?.[1]?.registerBlock &&
+        sent.transfer[1].lock.expiration.gte(sent.secret?.[1]?.registerBlock))
     )
       continue;
     const secrethash = key as Hash;
@@ -828,20 +855,14 @@ function autoExpire$(
 export const transferAutoExpireEpic = (
   action$: Observable<RaidenAction>,
   state$: Observable<RaidenState>,
+  { config$ }: RaidenEpicDeps,
 ): Observable<transferExpire.request | transfer.failure> =>
   action$.pipe(
     filter(isActionOf(newBlock)),
-    withLatestFrom(state$),
+    withLatestFrom(state$, config$),
     // exhaustMap ignores new blocks while previous request batch is still pending
-    exhaustMap(
-      ([
-        {
-          payload: { blockNumber },
-        },
-        state,
-      ]) => {
-        return autoExpire$(state, blockNumber, action$);
-      },
+    exhaustMap(([{ payload: { blockNumber } }, state, config]) =>
+      autoExpire$(action$, state, config, blockNumber),
     ),
   );
 
@@ -865,7 +886,13 @@ export const initQueuePendingEnvelopeMessagesEpic = (
       for (const [key, sent] of Object.entries(state.sent)) {
         const secrethash = key as Hash;
         // transfer already completed or channelClosed
-        if (sent.unlockProcessed || sent.lockExpiredProcessed || sent.channelClosed) continue;
+        if (
+          sent.unlockProcessed ||
+          sent.lockExpiredProcessed ||
+          sent.secret?.[1]?.registerBlock ||
+          sent.channelClosed
+        )
+          continue;
         // on init, request monitor presence of any pending transfer target
         yield matrixPresence.request(undefined, { address: sent.transfer[1].target });
         // Processed not received yet for LockedTransfer
@@ -967,7 +994,7 @@ const secretReveal$ = (
 ): Observable<transfer.failure | transferSecretReveal | messageSend.request> => {
   const request = action.payload.message;
   const secrethash = action.meta.secrethash;
-  if (!(secrethash in state.secrets)) {
+  if (!state.sent[secrethash]?.secret) {
     // shouldn't happen, as we're the initiator (for now), and always know the secret
     log.warn('SecretRequest for unknown secret', request, secrethash);
     return EMPTY;
@@ -1003,7 +1030,7 @@ const secretReveal$ = (
     const message: SecretReveal = {
       type: MessageType.SECRET_REVEAL,
       message_identifier: makeMessageId(),
-      secret: state.secrets[action.meta.secrethash].secret,
+      secret: state.sent[action.meta.secrethash].secret![1].value,
     };
     reveal$ = from(signMessage(signer, message, { log }));
   }
@@ -1078,6 +1105,7 @@ export const transferSecretRevealedEpic = (
         // don't unlock again if already unlocked, retry handled by transferUnlockedRetryMessageEpic
         // in the future, we may avoid retry until Processed, and [re]send once per SecretReveal
         state.sent[secrethash].unlock
+        // accepts secretReveal/unlock request even if registered on-chain
       )
         return;
       // transferSecret is noop if we already know the secret (e.g. we're the initiator)
@@ -1356,4 +1384,53 @@ export const withdrawSendConfirmationEpic = (
         },
       ),
     ),
+  );
+
+/**
+ * Monitors SecretRegistry and emits when a relevant secret gets registered
+ *
+ * @param action$ - Observable of RaidenActions
+ * @param state$ - Observable of RaidenStates
+ * @returns Observable of transferSecretRegistered actions
+ */
+export const monitorSecretRegistryEpic = (
+  {}: Observable<RaidenAction>,
+  state$: Observable<RaidenState>,
+  { secretRegistryContract }: RaidenEpicDeps,
+): Observable<transferSecretRegistered> =>
+  getEventsStream<[Hash, Secret, Event]>(secretRegistryContract, [
+    secretRegistryContract.filters.SecretRevealed(null, null),
+  ]).pipe(
+    withLatestFrom(state$),
+    filter(
+      ([[secrethash, , { blockNumber }], { sent }]) =>
+        // emits only if lock didn't expire yet
+        secrethash in sent && sent[secrethash].transfer[1].lock.expiration.gte(blockNumber!),
+    ),
+    map(([[secrethash, secret, event]]) =>
+      transferSecretRegistered(
+        {
+          secret,
+          txHash: event.transactionHash! as Hash,
+          txBlock: event.blockNumber!,
+          confirmed: undefined,
+        },
+        { secrethash },
+      ),
+    ),
+  );
+
+/**
+ * A simple epic to emit transfer.success when secret register is confirmed
+ *
+ * @param action$ - Observable of transferSecretRegistered actions
+ * @returns Observable of transfer.success actions
+ */
+export const transferSuccessOnSecretRegisteredEpic = (
+  action$: Observable<RaidenAction>,
+): Observable<transfer.success> =>
+  action$.pipe(
+    filter(transferSecretRegistered.is),
+    filter(action => !!action.payload.confirmed),
+    map(action => transfer.success({}, action.meta)),
   );