+++ raiden-ts/tests/e2e/raiden.spec.ts
@@ -209,6 +209,8 @@ describe('Raiden', () => {
                 ServiceRegistry: { address: partner as Address, block_number: 1 },
                 // eslint-disable-next-line @typescript-eslint/camelcase
                 UserDeposit: { address: partner as Address, block_number: 2 },
+                // eslint-disable-next-line @typescript-eslint/camelcase
+                SecretRegistry: { address: partner as Address, block_number: 3 },
               },
               address: accounts[1] as Address,
             },